brilloemulator: Use Android emulator instead of qemu
This CL adds an "emulator-arm" board to the BSP, and builds the
brilloemulator product on top of this BSP.

Two important changes in this CL are:

  * Including init.goldfish.rc
  * Including fstab.goldfish

The second is not strictly required, it could be named anything, but the
emulator searches for that file automatically. So it's nice to have.

The init file is required though. The emulator is called "goldfish" so
init will look for "init.goldfish.rc" first. We trampoline that to
"init.brillo.rc" after setting some goldfish specific values.

init.goldfish.sh is used to assign the emulator an IP address so that
ADB can work.

After building, use the following command to launch:

  emulator-arm -shell -no-window

This CL does not remove the vexpressa9 platform yet though it no longer
serves any purpose.

Change-Id: I921ca2c8e709c1c260258399f8b37760dbff7d83