brilloemulator: Do not use camera.goldfish.jpeg
camera.goldfish.jpeg is used to build the JpegCompressor
that is used in the emulated Camera2/3 HAL implementation.
The JpegCompressor has a dependency on the
Java framwork (libandroid_runtime) that we don't want on
Brillo.

This means that the emulated camera2/3 HAL implementations
won't be functional on the brilloemulator target.

BUG: 23161588
BUG: 22029972
BUG: 23090135

Change-Id: Id7475b0e873ee7e52f634364549f116ffabad8d9