Add default firewall rules.
This implements a default DROP policies with the necessary exceptions
for establishing connections, ICMP, and mDNS. We do this by installing
a script (init.brillo-setup-firewall.sh) into /system/etc and running it
once during init, prior to enabling firewalld.

This relies on a corresponding change to make the installed script
executable as a brillo service.

Bug: 23064386
Change-Id: I4a9f2b776d742f837aaefb356b750cf93d088e71