device/generic/brillo: Remove "<< EOF" syntax
It turns out that using "<< EOF" in /bin/sh write a
temporary file in /data/local.  This directory is
outside of the SELinux domain allowed to connectivity,
and will fail if enforcing is enabled, and produces an
error like this:

/system/bin/wifi_connect[288]: can't create temporary file /data/local/shn9pe3v.tmp: Permission denied

This CL switches the affected code to a function instead
of using the EOF syntax.

BUG: 22957017
Change-Id: If9d6b000ff27467880eb3e2bb31390a0bf9ae306
TEST: connectivity_test ap BrilloHello
TEST: Using a dragonbard-eng image