brillo: Add DTB support to the kernel build
bsp's can now set TARGET_KERNEL_DTB to the name
of a device tree in the kernel. This will be installed
as kernel-dtb in the product out folder.

If TARGET_KERNEL_DTB is not defined the build assumes
this kernel uses a concatenated dtb, which is the
default for Android and Brillo

This is required so the emulator can use a
seperate device tree.

BUG:23498693
Change-Id: Ib6afd32d4b3838d0c64a638063c184a1f0d58e23