brillo kernel: Add basic support for merging kernel configs
Some basic support for merging kernel configs. This allows
for a defconfig to be provided as a base and zero or more
config fragments to be added to it.

This doesn't handle things like config vars changing name
across kernel versions so knowlege of what you are
targetting is required

BUG: 23376873
Change-Id: I737424b04f5248bab138656bcaf331e4a47792ad