brilloemulator_arm: New brillo emulator target
Emulator target that builds the kernel in tree.
It kinda uses the new BSP model but as it has
no hardware in /hardware/bsp so we mock it up locally.
If we do add significant hardware, we should make
a new repo in /hardware/bsp for it.

Also installs a brilloemulator-arm script to the host
tools.

BUG: 23498693
Change-Id: I7b131abd2c1e1b3c2c64d710f68487673ef57554