device/generic/brillo: Create a WiFi Driver HAL
Create a modular resource for driver initialization and mode
change.  These functions are not necessary in the WiFi HAL,
which controls only run-time operation.  However, they are
critical for creating a uniform space for changing driver
modes, initial driver setup and querying the device names.

This removes the EAP prototype "wifi_init" script and
replaces it with a binary application that uses the HAL
hooks to setup the device.  The "wifi_connect" script is
now modified to remove all remaining references to the
"client_interface" variable which was provided by the
wifi_init scriptlet.  As a result, the "get_status"
method must iterate over all available devices to
determine connectivity state.

While here remove the "check_ip_status" function which
has no callers.

BUG: 23801071
Change-Id: I5da55abfbd30ec4d6c65fd8d75c02e56f389d9f7
TEST: Boot and ensure AP mode is running (with device-specific HAL)