device/generic/brillo/wifi_connect : Fix DNS setup
This addresses a few corner cases in DNS startup in wifi_connect.
The script now ensures each portion of the connectivity stack is
up and running.  It verifies link-level and IP connectivity, then
ensures that there is a running dnsmasq process, starting it if
it is not running.  This covers the corner case where the DHCP
process takes so long that the "wifi_connect" script has timed
out.  Since wifi_connect is called periodically to retrieve
status, this will ensure that dnsmasq is launched if a lease is
attained later.

While here, make it so the "wifi_connect" script now only emits
"Connected" if the system has attained an IP addresss.  This
cleans up race conditions where we may report connectivity when
the link has come up but DHCP has not completed.

Since making this change also re-vamps the way processes are
killed, add a log message if wpa_supplicant was not found to
send a HUP to.

BUG: 24129168
TEST: Could not reproduce the bug above, but approximated this
by modifying the DHCP lease timeout.

Change-Id: I5308ae376e8b06c61e89fa69f3d87568be79f4ae