Do not start connectivity on device startup
Now that shill and apmanager is ready, connectivity tool is not needed
anymore.  Disable it for now so that it doesn't start during device
startup.

TODO: remove it from the brillo_base.mk and the source code.  Create a
wifi device context in shill and update driver sepolicy file to use that
context instead of connectivity_device.

Bug: 24575517
TEST=Verify connectivity is not running when dragonboard started

Change-Id: I8f2e3a1ca5824b8a67b4e109d98df53ef6f24c76