brillo_base: add libstdc++.
Android normally pulls libstdc++ in via embedded.mk, which most (all?)
devices inherit from. We are not inheriting embedded.mk, and the build
system adds implicit libstdc++ linkage to some included modules, so
we need to add libstdc++ or they die.

This bug is currently only seen on user builds, and the effect is that
pretty much all of our brillo daemons fail to start, so init
continually cycles through trying to restart them and cleaning them up
when they die. I expect it's not seen on userdebug/eng builds because
some extra package that's included explicitly brings in libstdc++.

Bug: http://b/24499744
Change-Id: I1b6f0da7456e7e97e870374941f7f0af33843ffa