brilloemulator-arm: Update embedded pairing code to 'hello'
To illustrate that embedded code is not limited to numeric digits,
use 'hello' instead of '0000'.

Also uncommented API keys and URLs to use the value from the .conf file
instead of hardcoded defaults (which are going to be removed).

And fixed the model manifest to 'AAAAA' since the generic device
shouldn't use dragonboard's manifest.

BUG: 24536219
Change-Id: Ibe7705394ff6823973357678639c8db27c729625