crash_reporter: Add chown capability
crash_reporter gets called by the kernel as the root uid and gid.  It needs
to then switch to the user and gid of the crashing process to access files in
/proc/<pid> and copy them to temporary space, then switches back to root to
generate the minidump for crash_sender to access.  It thus needs the chown
capability to allow root to access the files put in temporary space that were
generated while running as the crashed user.

Bug: 24678424
Change-Id: I01b6ba956bea2a143a1e34da3fb3e49ac5ad1a85