brillo: Add helper for DEVICE_PACKAGES
This allows BSPs to add packages. There is no support
in the build system for a board to add packages.
This helper adds a DEVICE_PACKAGES list to CUSTOM_MODULES.

CUSTOM_MODULES just get appended to the packages list.
Later we can try and get support for this in /build
so this helper will not be needed

BUG: 24915140

Change-Id: I27a6b5e38e7308369f327eac1172545fafe16463