Rename 'init_brillo_domain' macro to 'brillo_domain'.
"init" in 'init_brillo_domain' does not refer to any initialization,
but rather to the fact that the macro is used for daemons started by
'init'. Consequently, the name "init_brillo_domain" doesn't really
make sense. What makes sense is to call the macro 'brillo_domain',
which describes what any SELinux domain for a Brillo service has
access to.

Bug: None
Change-Id: I65b09f5109739bb2efa132fc040f99b16f1fb0ad