brilloemulator: exec qemu at the end
Currently the shell might stick around even though the qemu command is the
last thing executed.  It's a waste of resources, but also if you happen to
edit the file while it's running, the shell parses the file in the middle
when qemu exits.  Just exec to avoid both issues.

While we're here, standardize the notice message.

Change-Id: I60f357d348a33b1e68d7fae31dadb10cdd6f4ee5