brilloemulator_x86/x86_64: Enable KVM if available
For some reason, x86 and x86_64 emulators never receive a response
to DHCPDISCOVER requests for eth0, as is evident from dhcpcd logs
("adb logcat dhcpcd:D *:S"):

 dhcpcd is running with reduced privileges
 status changed to Init
 status changed to Discover
 eth0: soliciting a DHCP lease
 eth0: sending DISCOVER (xid 0x270e47e6), next in 3.6 seconds
 eth0: sending DISCOVER (xid 0x270e47e6), next in 8.6 seconds
 eth0: sending DISCOVER (xid 0x270e47e6), next in 15.4 seconds
 eth0: sending DISCOVER (xid 0x270e47e6), next in 32.3 seconds

On Ubuntu, running the emulators with KVM enabled fixes this issue,
although it is unclear why. In any case, KVM is beneficial to
emulator performance, so just enable it if it is available on the
host system.

Change-Id: I8a6b2faab6c22e8a697741a3999371f0fe89c995
Signed-off-by: Yu Ning <yu.ning@intel.com>