add python to brillo eng builds [reland]
We need this to run autotest tests.

The fs_config hack is a bit ugly, but it's just enough to unblock the
autotest guys.  We'll want to investigate this on a larger scale.

The failures seen earlier on bots should be fixed:
(1) emerge is fixed by CL:201300
(2) various LFS issues are fixed in master by enh@
(3) pw gecos errors on arm are fixed in CL:200085

BUG=b:26861037
TEST=brilloemulator x86_64 eng boots and `python` works

Change-Id: I5027176f9066e565dcc8fed62397d300e72ab364