kernel: allow target configs to be multiple files again
Clean up kernel configs again, hopefully this one will stick. Avoids
the need for "realpath" calls in the Makefile (since the merge tool
should be the one handling it), records the sources used to build the
config (instead of copying them around), and makes the source and output
variables easier to read.

Change-Id: Ie2b6fc222077b20f8b8f931c4c3084c6e692b38e