brillo: UEFI boot loader kernel booting.
Booting the Android kernel image in a UEFI environment. This CL provides
capabilities for the Brillo UEFI boot loader to specify a bootable partition in
order to load and boot it. This will be used to toggle between booting
into different partitions in A/B implementation.

BUG=29072323
TEST=Built in assertions passing, manually testing kernel booting using qemu.
Ongoing unit test development.

Change-Id: I0bb50019a4bd2cb7bfc54460172b38ec50819a0a